(TIN-6) Unite all controllers related with user in one user controller
